Do not assert layout in KnownPanicsLint.
