+++ tests/ui/mir/static-by-value-slice.rs
@@ -1,0 +1,13 @@
+//! Regression test for #140332
+//! KnownPanicsLint used to assert ABI compatibility in the interpreter,
+//! which ICEs with unsized statics.
+
+static mut S: [i8] = ["Some thing"; 1];
+//~^ ERROR the size for values of type `[i8]` cannot be known
+//~| ERROR mismatched types
+//~| ERROR mismatched types
+
+fn main() {
+    assert_eq!(S, [0; 1]);
+    //~^ ERROR use of mutable static is unsafe
+}
