+++ tests/ui/mir/static-by-value-slice.stderr
@@ -1,0 +1,33 @@
+error[E0277]: the size for values of type `[i8]` cannot be known at compilation time
+  --> $DIR/static-by-value-slice.rs:5:1
+   |
+LL | static mut S: [i8] = ["Some thing"; 1];
+   | ^^^^^^^^^^^^^^^^^^ doesn't have a size known at compile-time
+   |
+   = help: the trait `Sized` is not implemented for `[i8]`
+   = note: statics and constants must have a statically known size
+
+error[E0308]: mismatched types
+  --> $DIR/static-by-value-slice.rs:5:23
+   |
+LL | static mut S: [i8] = ["Some thing"; 1];
+   |                       ^^^^^^^^^^^^ expected `i8`, found `&str`
+
+error[E0308]: mismatched types
+  --> $DIR/static-by-value-slice.rs:5:22
+   |
+LL | static mut S: [i8] = ["Some thing"; 1];
+   |                      ^^^^^^^^^^^^^^^^^ expected `[i8]`, found `[i8; 1]`
+
+error[E0133]: use of mutable static is unsafe and requires unsafe function or block
+  --> $DIR/static-by-value-slice.rs:11:16
+   |
+LL |     assert_eq!(S, [0; 1]);
+   |                ^ use of mutable static
+   |
+   = note: mutable statics can be mutated by multiple threads: aliasing violations or data races will cause undefined behavior
+
+error: aborting due to 4 previous errors
+
+Some errors have detailed explanations: E0133, E0277, E0308.
+For more information about an error, try `rustc --explain E0133`.
